Resolved the "no hits" issue
During data analysis, there was an issue where a participant would generate no "hits" despite their gaze fixating on an AOI. This was resolved. The issue stemmed from a flipped y-axis in the Dynamic AOI Toolkit, which is reversed from the Cartesian coordinate system used by the Pupil Labs software.